added touch input scheme for ios
Probably not the best possible way of implementing touch input support,
but it does make the game playable on mobile.

The input scheme is adaptive with regard to the direction that pacman
is going. For example, going left creates the following touch input
grid:

 __________
|      |   |
|  U   |   |
|______| R |
|      |   |
|  D   |   |
|______|___|

Similarly every other direction produces a corresponding "adaptive"
grid.

Tested it briefly with a mouse and seemed reasonable, testing on a
real device might be warranted.